+++ ncem/torch_models/non_linear_ncem/train.py
@@ -1,39 +1,17 @@
 
 import os
-import argparse
 import pytorch_lightning as pl
 from pytorch_lightning.callbacks import ModelCheckpoint
 from torch_data.datasets.hartmann import Hartmann
-from torch_models.non_linear_ncem import NonLinearNCEM
+from model import NonLinearNCEM
 from torch_geometric import loader
-from torch_geometric.data import Dataset, Data
-from torch.utils.data import DataLoader
+from torch_geometric.data import Data
 import numpy as np
-
-
-def parse_args():
-    """
-    Program arguments
-    """
-    parser = argparse.ArgumentParser(description="Train")
-    parser.add_argument("--data_path", default="../example_data/hartmann", help="data path")
-    parser.add_argument("--dataset", type=str, default="hartmann", help="dataset to load", choices=["hartmann"])
-    parser.add_argument("--init_model", default=None, help="initial model to load")
-    parser.add_argument("--batch_size", type=int, default=1, help="train batch size")
-    parser.add_argument("--num_workers", type=int, default=12, help="the num of training process")
-    parser = NonLinearNCEM.add_model_specific_args(parser)
-    parser = pl.Trainer.add_argparse_args(parser)
-    args = parser.parse_args()
-
-    arg_groups = {}
-    for group in parser._action_groups:
-        group_dict = {a.dest: getattr(args, a.dest, None) for a in group._group_actions}
-        arg_groups[group.title] = argparse.Namespace(**group_dict)
-    return args, arg_groups
+from utils.argparser import parse_args
 
 
 def main():
-    args, arg_groups = parse_args()
+    args, arg_groups = parse_args(NonLinearNCEM)
 
     # Checkpoint settings
     checkpoint_callback = ModelCheckpoint(
@@ -86,13 +64,12 @@ def main():
     else:
         raise NotImplementedError()
 
-    trainer: pl.Trainer = pl.Trainer.from_argparse_args(args, callbacks=[checkpoint_callback])
-
     model = NonLinearNCEM(
         in_channels=n_input_features,
         out_channels=n_output_features,
         **vars(arg_groups["NonLinearNCEM"]))
-
+    
+    trainer: pl.Trainer = pl.Trainer.from_argparse_args(args, callbacks=[checkpoint_callback])
     trainer.fit(model=model, train_dataloaders=train_dataloader, val_dataloaders=val_dataloader)
