+++ ncem/utils/init_weights.py
@@ -1,0 +1,10 @@
+import torch
+import torch.nn as nn
+import torch_geometric.nn as geom_nn
+
+
+def init_weights(m):
+    if isinstance(m, nn.Linear):
+        torch.nn.init.kaiming_normal_(m.weight, nonlinearity='relu')
+        if m.bias is not None:
+            m.bias.data.fill_(0.01)
